+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * changed the default session max age from two hours to 15 minutes (you can keep two hours by setting the `TrackerConfig.SessionMaxAge` option)
 * added page views to statistics (including relative views and growth)
 * added "today" to growth rates, if it's within period (visitors, sessions, bounces, views)
-* added average session duration per day
+* added average session duration per day and path
 * added session duration growth to analyzer
 * updated dependencies